🧪 PC3904 task: split oversized doc-write tests
