+++ .agentplane/tasks/202607092208-PC3904/pr/github-title.txt
@@ -1,0 +1,1 @@
+🚧 PC3904 task: Decompose oversized test suites for v0.6.22 [202607092208-PC3904]
