+++ .agentplane/tasks/202607092208-PC3904/pr/meta.json
@@ -1,0 +1,13 @@
+{
+  "base": "main",
+  "branch": "task/202607092208-PC3904/decompose-oversized-test-suites-for-v0-6-22",
+  "created_at": "2026-07-11T12:30:11.362Z",
+  "diffstat_sha256": "sha256:d507cbfad2734efed5face96721dd0eb119a3a33a079c4afa55e6a2e8ffd415c",
+  "last_verified_at": null,
+  "schema_version": 1,
+  "task_id": "202607092208-PC3904",
+  "updated_at": "2026-07-11T12:30:11.362Z",
+  "verify": {
+    "status": "skipped"
+  }
+}
