+++ .agentplane/tasks/202607092208-PC3904/pr/review.md
@@ -1,0 +1,39 @@
+# PR Review
+
+Created: 2026-07-11T12:30:11.362Z
+
+## Task
+
+- Task: `202607092208-PC3904`
+- Title: Decompose oversized test suites for v0.6.22
+- Status: DOING
+- Branch: `task/202607092208-PC3904/decompose-oversized-test-suites-for-v0-6-22`
+- Canonical task record: `.agentplane/tasks/202607092208-PC3904/README.md`
+
+## Verification
+
+- State: pending
+- Note: Not recorded yet.
+- Canonical workflow state lives in the task README.
+
+## Handoff Notes
+
+- No handoff notes recorded yet. Use `agentplane pr note ...` to append one.
+
+<!-- BEGIN AUTO SUMMARY -->
+<details>
+<summary>Raw evidence</summary>
+
+- Updated: 2026-07-11T12:30:11.362Z
+- Branch: task/202607092208-PC3904/decompose-oversized-test-suites-for-v0-6-22
+- Head: computed live by `agentplane pr check` / `agentplane integrate`
+
+```text
+ .../src/cli/run-cli.core.tasks.doc-write.test.ts   | 462 +-------------------
+ ...run-cli.core.tasks.doc-write.validation.test.ts | 476 +++++++++++++++++++++
+ scripts/oversized-test-baseline.json               |   8 +-
+ 3 files changed, 479 insertions(+), 467 deletions(-)
+```
+
+</details>
+<!-- END AUTO SUMMARY -->
