+++ scripts/oversized-test-baseline.json
@@ -2,8 +2,8 @@
   "schema_version": 2,
   "threshold_lines": 1000,
   "budgets": {
-    "max_entries": 11,
-    "max_total_lines": 12574,
+    "max_entries": 10,
+    "max_total_lines": 11424,
     "max_file_lines": 1300
   },
   "entries": [
@@ -18,10 +18,6 @@
     {
       "file": "packages/agentplane/src/cli/run-cli.core.route-decision.test.ts",
       "lines": 1171
-    },
-    {
-      "file": "packages/agentplane/src/cli/run-cli.core.tasks.doc-write.test.ts",
-      "lines": 1014
     },
     {
       "file": "packages/agentplane/src/cli/run-cli.core.tasks.normalize-migrate.test.ts",
